Download op decoupled
