+++ Assets/Package/Editor/Resources.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 88ec2acf9e8cc204b9f73738226be179
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
