+++ Assets/Package/Editor/Resources/ModelDownloader.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a2552ec4614c3f94f82e4ae961c83de7
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
